Add Azure Pipelines (#13)
Adds azure pipeline build testing on Ubuntu and Windows and fixes builds on both systems.

+++ shared/include/assets/shader.hpp
@@ -23,7 +23,7 @@ struct Shader {
         }
     }
 
-    unsigned int handle = -1;
+    unsigned int handle = static_cast<unsigned int>(-1);
     constexpr static auto nohandle = static_cast<decltype(handle)>(-1);
 };
